+++ packages/nitro/src/index.ts
@@ -17,9 +17,17 @@ export default {
       (config.plugins as Array<unknown>).push(workflowRollupPlugin());
     });
 
-    // Temporary workaround for debug unenv mock
+    // NOTE: Temporary workaround for debug unenv mock
     if (!nitro.options.workflow?._vite) {
       nitro.options.alias['debug'] ??= 'debug';
+    }
+
+    // NOTE: Externalize .nitro/workflow to prevent dev reloads
+    if (nitro.options.dev) {
+      nitro.options.externals ||= {};
+      nitro.options.externals.external ||= [];
+      const outDir = join(nitro.options.buildDir, 'workflow');
+      nitro.options.externals.external.push((id) => id.startsWith(outDir));
     }
 
     // Add tsConfig plugin