+++ workbench/hono/nitro.config.ts
@@ -1,11 +1,8 @@
-import { defineNitroConfig } from 'nitro/config';
+import { defineConfig } from 'nitro';
 
-export default defineNitroConfig({
+export default defineConfig({
   modules: ['workflow/nitro'],
-  handlers: [
-    {
-      route: '/api/**',
-      handler: './server.ts',
-    },
-  ],
+  routes: {
+    '/**': './src/index.ts',
+  },
 });