+++ workbench/nitro-v3/nitro.config.ts
@@ -1,5 +1,6 @@
-import { defineNitroConfig } from 'nitro/config';
+import { defineConfig } from 'nitro';
 
-export default defineNitroConfig({
+export default defineConfig({
   modules: ['workflow/nitro'],
+  serverDir: './',
 });