+++ packages/builders/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @workflow/builders
 
+## 4.0.1-beta.11
+
+### Patch Changes
+
+- Updated dependencies [2fde24e]
+- Updated dependencies [4b70739]
+  - @workflow/core@4.0.1-beta.13
+  - @workflow/errors@4.0.1-beta.5
+
 ## 4.0.1-beta.10
 
 ### Patch Changes