+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @workflow/core
 
+## 4.0.1-beta.13
+
+### Patch Changes
+
+- 2fde24e: Use inline sourcemaps to prevent SWC read import error
+- 4b70739: Require specifying runId when writing to stream
+- Updated dependencies [4b70739]
+  - @workflow/world-vercel@4.0.1-beta.9
+  - @workflow/world-local@5.0.0-beta.9
+  - @workflow/world@4.0.1-beta.6
+  - @workflow/errors@4.0.1-beta.5
+
 ## 4.0.1-beta.12
 
 ### Patch Changes