+++ packages/sveltekit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @workflow/sveltekit
 
+## 4.0.0-beta.9
+
+### Patch Changes
+
+- @workflow/builders@4.0.1-beta.11
+
 ## 4.0.0-beta.8
 
 ### Patch Changes